+++ services/web/server/tests/unit/with_dbs/02/test_projects_states_handlers.py
@@ -6,6 +6,7 @@
 # pylint: disable=unused-variable
 
 import asyncio
+import contextlib
 import time
 from collections.abc import AsyncIterator, Awaitable, Callable, Iterator
 from copy import deepcopy
@@ -280,6 +281,7 @@ async def test_share_project(
     share_rights: dict,
     project_db_cleaner,
     request_create_project: Callable[..., Awaitable[ProjectDict]],
+    exit_stack: contextlib.AsyncExitStack,
 ):
     # Use-case: the user shares some projects with a group
 
@@ -303,7 +305,10 @@ async def test_share_project(
 
     # get another user logged in now
     await log_client_in(
-        client, {"role": user_role.name}, enable_check=user_role != UserRole.ANONYMOUS
+        client,
+        {"role": user_role.name},
+        enable_check=user_role != UserRole.ANONYMOUS,
+        exit_stack=exit_stack,
     )
     if new_project:
         # user 2 can get the project if user 2 has read access
@@ -1272,6 +1277,7 @@ async def test_open_shared_project_2_users_locked(
     clean_redis_table: None,
     mock_dynamic_scheduler_rabbitmq: None,
     mocked_notifications_plugin: dict[str, mock.Mock],
+    exit_stack: contextlib.AsyncExitStack,
 ):
     # Use-case: user 1 opens a shared project, user 2 tries to open it as well
     mock_project_state_updated_handler = mocker.Mock()
@@ -1332,7 +1338,10 @@ async def test_open_shared_project_2_users_locked(
 
     # 2. create a separate client now and log in user2, try to open the same shared project
     user_2 = await log_client_in(
-        client_2, {"role": user_role.name}, enable_check=user_role != UserRole.ANONYMOUS
+        client_2,
+        {"role": user_role.name},
+        enable_check=user_role != UserRole.ANONYMOUS,
+        exit_stack=exit_stack,
     )
     await _connect_websocket(
         socketio_client_factory,
@@ -1454,6 +1463,7 @@ async def test_open_shared_project_at_same_time(
     clean_redis_table,
     mock_dynamic_scheduler_rabbitmq: None,
     mocked_notifications_plugin: dict[str, mock.Mock],
+    exit_stack: contextlib.AsyncExitStack,
 ):
     NUMBER_OF_ADDITIONAL_CLIENTS = 10
     # log client 1
@@ -1475,6 +1485,7 @@ async def test_open_shared_project_at_same_time(
             new_client,
             {"role": user_role.name},
             enable_check=user_role != UserRole.ANONYMOUS,
+            exit_stack=exit_stack,
         )
         client_id = client_session_id_factory()
         sio = await _connect_websocket(
